mysite/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-filter==24.2
django-ratelimit==4.1.0
djangorestframework==3.14.0
docutils==0.21.2
drf-spectacular==0.27.2
flake8==7.1.0
future==1.0.0
idna==3.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
peewee==3.17.1
pefile==2021.9.3
pewee==0.1.0.dev9
pillow==10.4.0
pycodestyle==2.12.0
pyflakes==3.2.0
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.19.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
